u16 amount sizes to save memory

+++ rust/src/config.rs
@@ -14,9 +14,9 @@ pub static CONFIG: Lazy<Config> = Lazy::new(|| {
 #[derive(Serialize, Deserialize)]
 pub struct Config {
     // Chip parameters
-    pub stack_size: i32,
-    pub big_blind: i32,
-    pub small_blind: i32,
+    pub stack_size: u16,
+    pub big_blind: u16,
+    pub small_blind: u16,
 
     // Abstraction
     pub bet_abstraction: Vec<Vec<f64>>,
